\name{ase-package}
\alias{ase-package}
\alias{ase}
\docType{package}
\title{
ASE - Applied Statistics and Econometrics (Luiss University)
}
\description{
Collection of useful functions for "Applied Statistics and Econometrics".
}
\details{
\tabular{ll}{
Package: \tab ase\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-05\cr
License: \tab MIT\cr
}
}
\author{
Giuseppe Ragusa <gragusa@luiss.it>

Maintainer: Giuseppe Ragusa <gragusa@luiss.it>
}

