% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase.R
\docType{data}
\name{attend}
\alias{attend}
\title{Data on test score and attendance.}
\format{A data frame with 680 rows and 11 variables}
\usage{
data(attend)
}
\description{
These data were collected by Professors Ronald Fisher and Carl Liedholm
during a term in which they both taught principles of microeconomics at
Michigan State University.
}
\details{
\itemize{
\item  attend  classes attended out of 32
\item  termGPA GPA for term
\item  priGPA  cumulative GPA prior to term
\item  ACT     ACT score
\item  final   final exam score
\item  atndrte percent classes attended
\item  hwrte   percent homework turned in
\item  frosh   =1 if freshman
\item  soph    =1 if sophomore
\item  missed  number of classes missed
\item  stndfnl (final - mean)/sd
}
}
\keyword{datasets}
