% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase.R
\docType{data}
\name{ceo2}
\alias{ceo2}
\title{CEO Compensation}
\format{A data frame with 177 rows and 9 variables}
\usage{
data(ceo2)
}
\description{
A dataset containing a random sample of CEO's compensation information.
Source: May 6, 1991 Issue of Businessweek
}
\details{
\itemize{
 
\item    salary. numeric: 1990 compensation, $1000s
\item       age. numeric: age of CEO in years
\item   college. numeric: =1 if attended college
\item      grad. numeric: =1 if attended graduate school
\item    comten. numeric: years with company
\item    ceoten. numeric: years as ceo with company
\item     sales. numeric: 1990 firm sales, millions
\item  profits.  numeric: 1990 profits, millions
\item    mktval. numeric: market value, end 1990, mills.
}
}
\keyword{datasets}
