% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase.R
\docType{data}
\name{Cigarettes}
\alias{Cigarettes}
\title{Cigarette Consumption Panel Data}
\format{A data frame containing 48 observations on 7 variables for 2 periods.}
\usage{
data("Cigarettes")
}
\description{
Panel data on cigarette consumption for the 48 continental US States from
1985-1995.
}
\details{
\itemize{
\item    state Factor indicating state.
\item    year Factor indicating year.
\item    cpi Consumer price index.
\item    population State population.
\item    packs Number of packs per capita.
\item    income State personal income (total, nominal).
\item    tax Average state, federal and average local excise taxes for fiscal year.
\item  price Average price during fiscal year, including sales tax.
\item    taxs Average excise taxes for fiscal year, including sales tax.
}
}
\keyword{datasets}
