% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.rob.R
\name{confint_rob}
\alias{confint_rob}
\title{Confidence Intervals for Model Parameters}
\usage{
confint_rob(object, parm, level = 0.95, type = c("HC1", "const", "HC",
  "HC0", "HC2", "HC3", "HC4", "HC4m", "HC5"))
}
\arguments{
\item{object}{a fitted model object.}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing, all
parameters are considered.}

\item{level}{the confidence level required.}

\item{type}{}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence
limits for each parameter. These will be labelled as (1-level)/2 and 1 -
(1-level)/2 in \% (by default 2.5\% and 97.5\%).
}
\description{
Computes confidence intervals for one or more parameters in a fitted
model. There is a default and a method for objects inheriting from class
"\code{lm}".
}
\details{
\code{confint_rob} assumes asymptotic normality of the estimators of the parameters for
the fitted models, but does not assume conditional homoskedasticity.
}
\author{
Giuseppe Ragusa
}
