% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yet_another_functions_up_to_date_24112018.R
\name{CountsToAI}
\alias{CountsToAI}
\title{Calculates allelic imbalance from merged counts over given replicates (ai(sum_reps(gene))).}
\usage{
CountsToAI(df, reps = NA)
}
\arguments{
\item{df}{A dataframe of genes/transcripts and parental counts for technical replicates in columns.}

\item{reps}{An optional parameter for a range op replicates for consideration (default = all replicates in df).}
}
\value{
mean(mean(m_1,...,m_6))_SNP / mean(mean(m_1+p_1,...,m_6+p6))_SNP.
}
\description{
Calculates allelic imbalance from merged counts over given replicates (ai(sum_reps(gene))).
}
\examples{

}
