% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase.R
\docType{data}
\name{countycrime}
\alias{countycrime}
\title{County crime data}
\format{A data frame containing 630 observation on 26 variables.}
\source{
From C. Cornwell and W. Trumball (1994), “Estimating the
Economic Model of Crime with Panel Data,” Review of Economics and Statistics 76, 360-366.
}
\usage{
data("countycrime")
}
\description{
Panel data on county level crime data.
}
\details{
\itemize{
\item     county               county identifier
\item     year                        81 to 87
\item     crmrte     crimes committed per person
\item     prbarr         'probability' of arrest
\item     prbconv     'probability' of conviction
\item     prbpris 'probability' of prison sentenc
\item     avgsen             avg. sentence, days
\item     polpc               police per capita
\item     density             people per sq. mile
\item     taxpc          tax revenue per capita
\item      west           =1 if in western N.C.
\item   central           =1 if in central N.C.
\item     urban                   =1 if in SMSA
\item  pctmin80            perc. minority, 1980
\item      wcon       weekly wage, construction
\item      wtuc    wkly wge, trns, util, commun
\item      wtrd wkly wge, whlesle, retail trade
\item      wfir    wkly wge, fin, ins, real est
\item      wser      wkly wge, service industry
\item      wmfg         wkly wge, manufacturing
\item      wfed         wkly wge, fed employees
\item      wsta       wkly wge, state employees
\item      wloc        wkly wge, local gov emps
\item      mix offense mix: face-to-face/other
\item      pctymle              percent young male
}
}
\keyword{datasets}
