% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yet_another_functions_up_to_date_24112018.R
\name{CreateDeltaAIForAPairRepsDF}
\alias{CreateDeltaAIForAPairRepsDF}
\title{Creates a tab with gene mean coverage and AI deltas for a pair of technical replicates.}
\usage{
CreateDeltaAIForAPairRepsDF(df, reps, thr = 0, thrUP = F)
}
\arguments{
\item{reps}{A parameter for setting 2 replicates for consideration.}

\item{thr}{Threshold for min gene coverage (default = 0).}

\item{thrUP}{Threshold for max gene coverage (default = F).}

\item{tab}{A dataframe of genes/transcripts and parental counts for technical replicates in columns.}
}
\value{
A tab with gene mean coverage and AI deltas for a pair of technical replicates.
}
\description{
Creates a tab with gene mean coverage and AI deltas for a pair of technical replicates.
}
\examples{

}
