% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yet_another_functions_up_to_date_24112018.R
\name{CreateObservedQuartilesDF}
\alias{CreateObservedQuartilesDF}
\title{Creates a table of parwise AI differences quartilles info for techreps in given table, binned to log intervals.}
\usage{
CreateObservedQuartilesDF(df, P, ep, logbase = T, coverageLimit,
  group = "")
}
\arguments{
\item{df}{A dataframe-output of CreateMergedDeltaAIPairwiseDF().}

\item{P}{A vector of %-quartiles.}

\item{ep}{The log base for binning (0^b, 1^b, ...).}

\item{coverageLimit}{Gene coverage limit for consideration.}

\item{group}{An ptional distinguishible name (default = '').}

\item{logbare}{The binary parameter if we deal with log base (default = T).}
}
\value{
A table of parwise AI differences info for techreps in given table, binned to log intervals.
}
\description{
Creates a table of parwise AI differences quartilles info for techreps in given table, binned to log intervals.
}
\examples{

}
