% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ase.R
\docType{data}
\name{crimenc}
\alias{crimenc}
\title{Crime and punishment in North Carolina}
\format{A data frame with 630 rows and 25 variables}
\usage{
data(crimenc)
}
\description{
A panel of 90 observations from 1981 to 1987 about crime and punishment in
Noth Caroline.
}
\details{
\itemize{
\item county county identifier
\item year year from 1981 to 1987
\item crmrte. crimes committed per person
\item prbarr. 'probability' of arrest
\item prbconv. 'probability' of conviction
\item prbpris. 'probability' of prison sentenc
\item avgsen. average sentence, days
\item polpc. police per capita
\item density. people per square mile
\item taxpc. tax revenue per capita
\item region. one of 'other', 'west' or 'central'
\item smsa. 'yes' or 'no' if in SMSA
\item pctmin. percentage minority in 1980
\item wcon. weekly wage in construction
\item wtuc. weekly wage in trns, util, commun
\item wtrd. weekly wage in whole sales and retail trade
\item wfir. weekly wage in finance, insurance and real estate
\item wser. weekly wage in service industry
\item wmfg. weekly wage in manufacturing
\item wfed. weekly wage of federal emplyees
\item wsta. weekly wage of state employees
\item wloc. weekly wage of local governments employee
\item mix. offence mix: face-to-face/other
\item pctymle. percentage of young males
}
}
\keyword{datasets}
