% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yet_another_functions_up_to_date_24112018.R
\name{FitLmIntercept}
\alias{FitLmIntercept}
\title{Fits linear model to logarithmic data and counts intersept for model with *1/2 restriction.}
\usage{
FitLmIntercept(inDf, binNObs, morethan = 10, logoutput = T)
}
\arguments{
\item{inDf}{A dataframe-output of CreateObservedQuartilesDF().}

\item{morethan}{Theshold on gene coverage for lm (default = 10).}

\item{logoutput}{Return log intercept? (default = true)}

\item{N_obs_bin}{Threshold on number of observations per bin.}
}
\value{
lm intercept or log2(lm intercept).
}
\description{
Fits linear model to logarithmic data and counts intersept for model with *1/2 restriction.
}
\examples{

}
