% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yet_another_functions_up_to_date_24112018.R
\name{GetGatkPipelineTabs}
\alias{GetGatkPipelineTabs}
\title{(GATK pipeline) Concatenate vertically (uniting merge) tables from inFiles.}
\usage{
GetGatkPipelineTabs(inFiles, nReps)
}
\arguments{
\item{inFiles}{A vector of pathes to files.}

\item{nReps}{A vector of numbers, each -- number of replicates in corresponding file.}
}
\value{
A techreps-concatenated table with means/counts, rows corresponds to genes.
}
\description{
(GATK pipeline) Concatenate vertically (uniting merge) tables from inFiles.
}
\examples{

}
