% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivreg.R
\name{ivreg}
\alias{ivreg}
\title{ivreg}
\usage{
ivreg(formula, instruments, data, subset, weights, na.action,
  contrasts = NULL, method = "tsls", start = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted. The details of
model specification are given under ‘Details’.}

\item{instruments}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the instruments to be used.}

\item{data}{an optional data frame, list or environment (or object
coercible by as.data.frame to a data frame) containing the variables in the
model. If not found in data, the variables are taken from
environment(formula), typically the environment from which lm is called.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{weights}{an optional vector of weights to be used in the fitting
process. Should be NULL or a numeric vector. If non-NULL, weighted least
squares is used with weights weights (that is, minimizing sum(w*e^2));
otherwise ordinary least squares is used. See also ‘Details’,}

\item{na.action}{a function which indicates what should happen when the
data contain NAs. The default is set by the na.action setting of options, and
is na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another
possible value is NULL, no action. Value na.exclude can be useful.}

\item{contrasts}{an optional list. See the contrasts.arg of model.matrix.default.}

\item{method}{the method to be used.}

\item{start}{initial values. Needed only when methods is different than "tsls" or "gmm"}

\item{...}{other arguments}
}
\value{
\code{NULL}
}
\description{
Estimate IV models
}
\details{
Details
}
