% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yet_another_functions_up_to_date_24112018.R
\name{MeanCoverage}
\alias{MeanCoverage}
\title{Calculates mean coverage mat+pat among given replicates.}
\usage{
MeanCoverage(df, reps = NA)
}
\arguments{
\item{df}{A dataframe of genes/transcripts and parental counts for technical replicates in columns.}

\item{reps}{An optional parameter for a range op replicates for consideration (default = all replicates in df).}
}
\value{
Mean coverage mat+pat among given replicates.
}
\description{
Calculates mean coverage mat+pat among given replicates.
}
\examples{

}
