% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{payroll}
\alias{payroll}
\title{SURVEY ON HOUSEHOLD INCOME AND WEALTH data}
\format{A data frame with 5700 rows and 5 variables}
\usage{
data(payroll)
}
\description{
A dataset contains data about payroll workers from the SWIG (Bank of Italy).
}
\details{
\itemize{
  \item wage. numeric: wages for payroll workers
  \item educ. numeric: years of education
  \item educf. factor: education levels.
  \item gender. factor: gender
  \item age. numeric: age of respondent
}
}
\keyword{datasets}

