% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary_rob.R
\name{print.summary_rob}
\alias{print.summary_rob}
\title{Printing Summary}
\usage{
\method{print}{summary_rob}(x, digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor, signif.stars = FALSE, call = FALSE, ...)
}
\description{
\code{print.summary_rob} is the print method for \code{summary_rob} model summaries.
}
\details{
\code{print.summary_rob} is the print method for \code{summary_rob} model summaries.
}
\author{
Giuseppe Ragusa
}
