% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary_rob}
\alias{summary_rob}
\alias{summary_rob.lm}
\alias{summary_rob.plm}
\title{Object Summaries}
\usage{
summary_rob(obj, ...)

\method{summary_rob}{lm}(object, alpha = 0.05, type = c("HC1", "const",
  "HC", "HC0", "HC2", "HC3", "HC4", "HC4m", "HC5"))

\method{summary_rob}{plm}(obj, alpha = 0.05, cluster = FALSE,
  type = c("HC1", "const", "HC", "HC0", "HC2", "HC3", "HC4", "HC4m", "HC5"))
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{alpha}{significance level for the Wald test}

\item{type}{the type of heteroskedastic robust variance estimator}
}
\value{
It depends

\code{NULL}

\code{NULL}
}
\description{
\code{summary_rob} produces result summaries of the results of various model
fitting functions using as default heteroskedastic robust standard errors.
}
\author{
Giuseppe Ragusa
}

