% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rob.R
\name{summary_rob}
\alias{summary_rob}
\alias{summary_rob.lm}
\alias{summary_rob.glm}
\alias{summary_rob.plm}
\title{Summarize regression objects}
\usage{
summary_rob(object, ...)

\method{summary_rob}{lm}(object, alpha = 0.05, type = c("HC1", "const",
  "HC", "HC0", "HC2", "HC3", "HC4", "HC4m", "HC5"), omit_factor = NULL)

\method{summary_rob}{glm}(object, alpha = 0.05, type = c("HC1", "const",
  "HC", "HC0", "HC2", "HC3", "HC4", "HC4m", "HC5"))

\method{summary_rob}{plm}(object, alpha = 0.05, cluster = FALSE,
  type = c("HC1", "const", "HC", "HC0", "HC2", "HC3", "HC4", "HC4m", "HC5"))
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{alpha}{significance level for the Wald test.}

\item{type}{the type of heteroskedastic robust variance estimator.}

\item{omit_factor}{whether to omit factor in output.}

\item{cluster}{for the plm method cluster standard errors at the individual level.}
}
\value{
A \code{summary_rob} object.
}
\description{
\code{summary_rob} produces result summaries of the results of various model
fitting functions using as default heteroskedastic robust standard errors.
}
\author{
Giuseppe Ragusa
}
